requirements.txt
# Core dependencies
torch>=2.0.0
torch-geometric>=2.3.0
rdkit>=2023.3.1
numpy>=1.24.0
scipy>=1.10.0
pandas>=2.0.0
tqdm>=4.65.0
wandb>=0.15.0
hydra-core>=1.3.0
omegaconf>=2.3.0
scikit-learn>=1.2.0
matplotlib>=3.7.0
seaborn>=0.12.0

# For molecular processing
mordred>=1.2.0
moses>=0.3.0

# Additional utilities
pyyaml>=6.0
tensorboard>=2.13.0
jupyterlab>=4.0.0